refactor(activesupport): move String#succ to the string core-ext, match Rails' TWZ fixture
Review round 1.

- `stringSucc` -> `core-ext/string/succ.ts` as `succ`, per the story's
  "Converged shape" note. It is Ruby core `String#succ`, so there is no
  `core_ext/string/*.rb` counterpart — tagged `@noRailsEquivalent PERMANENT` —
  but the string core-ext is where it belongs rather than beside the range
  files that consume it. `range-ext.ts` novel: 4 -> 3.
- The two `*_on_time_with_zone` tests built `makeRange(twz, twz)`; Rails uses
  `(twz - 1.hour)..twz` (range_ext_test.rb:252,259). Now
  `makeRange(twz.minus(hours(1)), twz)`.
